playlist https://www.youtube.com/playlist?list=PLu13hPnba_XCuinhx0D3H_O6Z6yHhOtp-
video 4

issues 15 Create module 'dlog_hero'

alfa version

Plugin work, but there is a problem with the cache. If you go to the page of the entity and then to the main page (or vice versa), the second plugin is not included. In method __construct parameter "plugin_type" (required plugin type "entity" or "path") is pulled from backend cache.